+++ homepage/example.js
@@ -124,17 +124,17 @@ function makePanel(e) {
  * Measure the display size
  *
  */
-function measureDisplaySize(e) {
-  RemoteCalibrator.displaySize(displayData => {
-    printMessage(
-      `Display size is ${displayData.value.displayWidthPx} px in width and ${
-        displayData.value.displayHeightPx
-      } px in height, measured at ${parseTimestamp(displayData.timestamp)}.`
-    )
+// function measureDisplaySize(e) {
+//   RemoteCalibrator.displaySize(displayData => {
+//     printMessage(
+//       `Display size is ${displayData.value.displayWidthPx} px in width and ${
+//         displayData.value.displayHeightPx
+//       } px in height, measured at ${parseTimestamp(displayData.timestamp)}.`
+//     )
 
-    changeClass(e.target, 'complete')
-  })
-}
+//     changeClass(e.target, 'complete')
+//   })
+// }
 
 /**
  *
@@ -398,10 +398,10 @@ function getGazeNow() {
  * Get environment info
  *
  */
-function getEnvironment(e) {
-  RemoteCalibrator.environment(data => {
-    printMessage('Environment: ' + data.value.description + '.')
+// function getEnvironment(e) {
+//   RemoteCalibrator.environment(data => {
+//     printMessage('Environment: ' + data.value.description + '.')
 
-    changeClass(e.target, 'complete')
-  })
-}
+//     changeClass(e.target, 'complete')
+//   })
+// }